chore: refactor plotting
