+++ examples/trades.py
@@ -7,7 +7,7 @@ from matplotlib.figure import Figure
 
 from vegapy.service.service import Service
 from vegapy.service.networks.constants import Network
-from vegapy.plot.trades import (
+from vegapy.visualisations.overlay import (
     overlay_size,
     overlay_price,
     overlay_volume,
@@ -72,6 +72,11 @@ PARSER.add_argument(
 )
 PARSER.add_argument(
     "--infrastructure_fee",
+    action="store_true",
+    help="Overlay cumulative infrastructure fees generated by trades on plot.",
+)
+PARSER.add_argument(
+    "--liquidations",
     action="store_true",
     help="Overlay cumulative infrastructure fees generated by trades on plot.",
 )
@@ -148,6 +153,13 @@ if __name__ == "__main__":
             axl,
             trades=trades,
             asset_decimals=asset.details.decimals,
+            cumulative=True,
+        )
+    if args.liquidations:
+        overlay_liquidations(
+            axl,
+            trades=trades,
+            size_decimals=market.position_decimal_places,
             cumulative=True,
         )
